resolves the dependencies methods that may not exist in a class
